+++ Forms/CommandLineForm.vb
@@ -14,7 +14,7 @@ Public Class CommandLineForm
 
     Public Sub New(params As CommandLineParams)
         InitializeComponent()
-        SimpleUI.ScaleClientSize(37, 25)
+        SimpleUI.ScaleClientSize(37, 26)
         rtbCommandLine.ScrollBars = RichTextBoxScrollBars.None
         Dim singleList As New List(Of String)